refactoring memo editor
